refactor: Add verifyTicketInformations helper for activities service
